+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Release 0.11.0 (unreleased)
 * Add more tests and documentation for patching (#888)
 * Restrict ``src`` to string only in schema (#889)
 * Don't consider ignored files for determining local changes (#350)
+* Switch from pykwalify to yamale for yaml validation
 
 Release 0.10.0 (released 2025-03-12)
 ====================================